refactor: separate snapshot UI testing by component
- Split component snapshots from combined index.test.tsx into individual test files
- Created separate test files: ChatInterface.test.tsx, CodeEditor.test.tsx, ComposerApp.test.tsx, GitWorktreeCreatePrompt.test.tsx, Layout.test.tsx, Sidebar.test.tsx, TelemetryConsentPrompt.test.tsx
- Generated individual snapshot files for each component
- Updated index.test.tsx to only contain GitWorktreeCli tests
- Removed old combined snapshot file

This improves test organization and maintainability by isolating component snapshots.
